requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cv==1.0.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastremap==1.16.1
filelock==3.18.0
fill_voids==2.0.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grad-cam==1.5.5
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.21
langchain-core==0.3.49
langchain-groq==0.3.1
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
lazy_loader==0.4
llvmlite==0.44.0
lungmask @ git+https://github.com/JoHof/lungmask.git@0653c7019200eb27253c540f731583466fc3c304
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.2.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.6
numba==0.61.2
numpy==2.2.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pydicom==3.0.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
setuptools==80.9.0
shap==0.48.0
simpleitk==2.5.0
simsimd==6.4.9
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.43.2
stringzilla==3.12.5
sudoku-solver==0.2
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.6.11
timm==1.0.15
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
ttach==0.0.3
twilio==9.5.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
